arquivo test ct004-01, ct004-02

+++ src/main/java/aula/test/Calculadora.java
@@ -13,7 +13,7 @@ public class Calculadora {
 	public int mutiplicar(int valor1, int valor2) {
 		return valor1 * valor2;
 	}
-	public double dividir(double valor1, double valor2) {
+	public double dividir(int valor1, int valor2) {
 		return valor1 / valor2;
 	}
